requirements.txt
aniso8601==9.0.1
click==8.1.3
Flask==2.1.2
Flask-JWT-Extended==4.4.1
Flask-Pydantic==0.9.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pydantic==1.9.1
PyJWT==2.4.0
PyMySQL==1.0.2
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.37
typing_extensions==4.2.0
Werkzeug==2.1.2
zipp==3.8.0
